pyproject.toml
[project]
name = "async-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.26.0",
    "prompt-toolkit>=3.0.52",
    "smolagents>=1.24.0",
    "tiktoken>=0.12.0",
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
    "pytest-asyncio>=1.3.0",
    "pyte>=0.8.2",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.uv.workspace]
members = ["my_project"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
